base/requirements.txt
Flask-SQLAlchemy
Flask-Script
Flask-Security
MySQL-python
